[PAB-476] Refactor extraction tests
+++ Moved local debug helper "tests" to more appropriate location
--- Removed redundant test file
