+++ tests/extraction_tests/extraction_tests.py
@@ -1,7 +1,12 @@
 import pandas as pd
+import pytest
 from pandas import Timestamp
 
-from core.extraction.towns_fund import extract_submission_details
+# isort: off
+from core.extraction.towns_fund import extract_submission_details, ingest_towns_fund_data
+from core.extraction.towns_fund_round_two import ingest_round_two_data_towns_fund
+
+# isort: on
 
 
 def test_submission_extract():
@@ -51,3 +56,25 @@ def test_submission_extract():
     ]
 
     assert test_output["Reporting Round"] == ["1", "2", "3", "4", "5", "6", "7", "8", "9"]
+
+
+# Test intended only as a local debug tool
+@pytest.mark.skip(reason="currently this is just a pytest/pycharm debug entrypoint for ingest work")
+def test_ingest_towns_fund_template():
+    towns_fund_data = pd.read_excel(
+        "EXAMPLE_TF_Reporting_Template_-_TD_-_Newhaven_-_DDMMYY.xlsx",
+        sheet_name=None,  # extract from all sheets
+    )
+    ingest_towns_fund_data(towns_fund_data)
+
+
+# Test intended only as a local debug tool
+@pytest.mark.skip(reason="currently this is just a pytest/pycharm debug entrypoint for ingest work")
+def test_ingest_round_two_historical():
+    # TODO: currently testing with small subset of data (to allow reasonable debugging speed)
+    round_two_data = pd.read_excel(
+        "Round 2 Reporting - Consolidation (MASTER).xlsx",
+        # "Round 2 Reporting - Consolidation.xlsx",
+        sheet_name=None,  # extract from all sheets
+    )
+    ingest_round_two_data_towns_fund(round_two_data)
